requirements.txt
# Base
Django==3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.1
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.4.0
requests==2.24.0

# Linting
flake8==3.8.3
flake8-django==1.1.1
isort==5.4.2

# Testing
coverage==5.2.1
django-nose==1.4.6